pyproject.toml
[tool.poetry]
name = "whitehathacker"
version = "0.1.0"
description = ""
authors = ["Toshiya Takeshita <toshi@atbat.jp>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
